requirements.txt
accounts==0.2.4
channels==4.0.0
core==1.0.1
Django==4.1.1
djangorestframework==3.14.0
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
scikit_learn==1.2.2
seaborn==0.12.2
